+++ CHANGELOG.md
@@ -18,6 +18,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Lazily load behaviors as separate Webpack chunks.
 - Avoid referencing named behaviors via window.farmOS.map since this breaks when farmOS-map is bundled e.g. in FieldKit.
 - Honor weights in async behavior attaching and expose `instance.defaultBehaviorsAttached`.
+- Use `save-remote-file-webpack-plugin` to download `ol.js` and `ol.css`.
 
 ### Fixed